backend/requirements2.txt

Flask
Flask-Cors #flask-cors
requests
beautifulsoup4
newspaper3k
lxml_html_clean
pymongo
#torch